requirements.txt
ipywidgets
jupyterlab
matplotlib
notebook
numpy
PyYAML
scipy
# torch
# torchaudio
# torchvision
torchlpc
torchcomp
torch_fftconv
torchlibrosa
kazane
soundfile
librosa
wandb
pingouin
hydra-core
pyloudnorm
auraloss
lightning[pytorch-extra]
tqdm